requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
et-xmlfile==1.1.0
filelock==3.12.4
fsspec==2023.9.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2
numba==0.58.1
numpy==1.26.0
openai-whisper==20230918
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12
tiktoken==0.3.3
torch==2.1.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
webdriver-manager==4.0.1
wsproto==1.2.0
